@@ -24,9 +24,9 @@ install_requires =
     jinja2~=3.0
     pyyaml~=6.0
     requests
-    snowflake-connector-python>=2.8,<4.0
+    snowflake-connector-python>=2.8,<5.0
     structlog~=24.1.0
-python_requires = >=3.8
+python_requires = >=3.10
 include_package_data = True
 
 [options.entry_points]
