@@ -43,13 +43,7 @@ class ChangeHistoryTable:
                 raise ValueError(f"Invalid change history table name: {table_str}")
 
         return cls(
-            table_name=get_snowflake_identifier_string(
-                input_value=table_name, input_type="table_name"
-            ),
-            schema_name=get_snowflake_identifier_string(
-                input_value=schema_name, input_type="schema_name"
-            ),
-            database_name=get_snowflake_identifier_string(
-                input_value=database_name, input_type="database_name"
-            ),
+            table_name=get_snowflake_identifier_string(input_value=table_name, input_type="table_name"),
+            schema_name=get_snowflake_identifier_string(input_value=schema_name, input_type="schema_name"),
+            database_name=get_snowflake_identifier_string(input_value=database_name, input_type="database_name"),
         )
