@@ -5,12 +5,12 @@ from pathlib import Path
 import pytest
 
 from schemachange.session.Script import (
+    AlwaysScript,
+    RepeatableScript,
     Script,
     VersionedScript,
-    RepeatableScript,
-    AlwaysScript,
-    script_factory,
     get_all_scripts_recursively,
+    script_factory,
 )
 
 
@@ -154,9 +154,7 @@ class TestScript:
         file_path = Path("nested/file/V__something.sql.jinja")
         with pytest.raises(ValueError) as e:
             script_factory(file_path)
-        assert str(file_path) in str(
-            e.value
-        ) and "Versioned migrations must be prefixed with a version" in str(e.value)
+        assert str(file_path) in str(e.value) and "Versioned migrations must be prefixed with a version" in str(e.value)
 
 
 class TestGetAllScriptsRecursively:
@@ -164,7 +162,7 @@ class TestGetAllScriptsRecursively:
         root_directory = Path("some_path")
         result = get_all_scripts_recursively(root_directory)
 
-        assert result == dict()
+        assert result == {}
 
     def test_given_just_non_change_files_should_return_empty(self, fs):
         fs.create_file(Path("scripts") / "README.txt")
@@ -172,7 +170,7 @@ class TestGetAllScriptsRecursively:
         fs.create_file(Path("scripts") / "subfolder" / "subfolder2" / "testing.py")
         result = get_all_scripts_recursively(Path("scripts"))
 
-        assert result == dict()
+        assert result == {}
 
     ############################
     #### Version file tests ####
@@ -181,9 +179,7 @@ class TestGetAllScriptsRecursively:
     def test_version_number_regex_numeric_happy_path(self, fs):
         fs.create_file(Path("scripts") / "V1.1.1__initial.sql")
         fs.create_file(Path("scripts") / "subfolder" / "V1.1.2__update.SQL")
-        fs.create_file(
-            Path("scripts") / "subfolder" / "subfolder2" / "V1.1.3__update.sql"
-        )
+        fs.create_file(Path("scripts") / "subfolder" / "subfolder2" / "V1.1.3__update.sql")
 
         result = get_all_scripts_recursively(
             Path("scripts"),
@@ -202,9 +198,7 @@ class TestGetAllScriptsRecursively:
                 Path("scripts"),
                 version_number_regex=r"\d\.\d\.\d",  # noqa: W605
             )
-        assert str(e.value).startswith(
-            "change script version doesn't match the supplied regular expression"
-        )
+        assert str(e.value).startswith("change script version doesn't match the supplied regular expression")
 
     def test_version_number_regex_text_happy_path(self, fs):
         fs.create_file(Path("scripts") / "Va.b.c__initial.sql")
@@ -223,16 +217,12 @@ class TestGetAllScriptsRecursively:
                 Path("scripts"),
                 version_number_regex=r"[a-z]\.[a-z]\.[a-z]",  # noqa: W605
             )
-        assert str(e.value).startswith(
-            "change script version doesn't match the supplied regular expression"
-        )
+        assert str(e.value).startswith("change script version doesn't match the supplied regular expression")
 
     def test_given_version_files_should_return_version_files(self, fs):
         fs.create_file(Path("scripts") / "V1.1.1__initial.sql")
         fs.create_file(Path("scripts") / "subfolder" / "V1.1.2__update.SQL")
-        fs.create_file(
-            Path("scripts") / "subfolder" / "subfolder2" / "V1.1.3__update.sql"
-        )
+        fs.create_file(Path("scripts") / "subfolder" / "subfolder2" / "V1.1.3__update.sql")
         result = get_all_scripts_recursively(Path("scripts"))
 
         assert len(result) == 3
@@ -243,15 +233,11 @@ class TestGetAllScriptsRecursively:
     def test_given_same_version_twice_should_raise_exception(self, fs):
         fs.create_file(Path("scripts") / "V1.1.1__initial.sql")
         fs.create_file(Path("scripts") / "subfolder" / "V1.1.1__update.sql")
-        fs.create_file(
-            Path("scripts") / "subfolder" / "subfolder2" / "V1.1.2__update.sql"
-        )
+        fs.create_file(Path("scripts") / "subfolder" / "subfolder2" / "V1.1.2__update.sql")
 
         with pytest.raises(ValueError) as e:
             get_all_scripts_recursively(Path("scripts"))
-        assert str(e.value).startswith(
-            "The script version 1.1.1 exists more than once (second instance"
-        )
+        assert str(e.value).startswith("The script version 1.1.1 exists more than once (second instance")
 
     def test_given_single_version_file_should_extract_attributes(self, fs):
         fs.create_file(Path("scripts") / "subfolder" / "V1.1.1.1__THIS_is_my_test.sql")
@@ -261,42 +247,30 @@ class TestGetAllScriptsRecursively:
         assert len(result) == 1
         script = result["v1.1.1.1__this_is_my_test.sql"]
         assert script.name == "V1.1.1.1__THIS_is_my_test.sql"
-        assert (
-            script.file_path
-            == Path("scripts") / "subfolder" / "V1.1.1.1__THIS_is_my_test.sql"
-        )
+        assert script.file_path == Path("scripts") / "subfolder" / "V1.1.1.1__THIS_is_my_test.sql"
         assert script.type == "V"
         assert script.version == "1.1.1.1"
         assert script.description == "This is my test"
 
     def test_given_single_version_jinja_file_should_extract_attributes(self, fs):
-        fs.create_file(
-            Path("scripts") / "subfolder" / "V1.1.1.2__THIS_is_my_test.sql.jinja"
-        )
+        fs.create_file(Path("scripts") / "subfolder" / "V1.1.1.2__THIS_is_my_test.sql.jinja")
 
         result = get_all_scripts_recursively(Path("scripts"))
 
         assert len(result) == 1
         script = result["v1.1.1.2__this_is_my_test.sql"]
         assert script.name == "V1.1.1.2__THIS_is_my_test.sql"
-        assert (
-            script.file_path
-            == Path("scripts") / "subfolder" / "V1.1.1.2__THIS_is_my_test.sql.jinja"
-        )
+        assert script.file_path == Path("scripts") / "subfolder" / "V1.1.1.2__THIS_is_my_test.sql.jinja"
         assert script.type == "V"
         assert script.version == "1.1.1.2"
         assert script.description == "This is my test"
 
-    def test_given_same_version_file_with_and_without_jinja_extension_should_raise_exception(
-        self, fs
-    ):
+    def test_given_same_version_file_with_and_without_jinja_extension_should_raise_exception(self, fs):
         fs.create_file(Path("scripts") / "V1.1.1__initial.sql")
         fs.create_file(Path("scripts") / "V1.1.1__initial.sql.jinja")
         with pytest.raises(ValueError) as e:
             get_all_scripts_recursively(Path("scripts"))
-        assert str(e.value).startswith(
-            "The script name V1.1.1__initial.sql exists more than once (first_instance"
-        )
+        assert str(e.value).startswith("The script name V1.1.1__initial.sql exists more than once (first_instance")
 
     ###########################
     #### Always file tests ####
@@ -321,9 +295,7 @@ class TestGetAllScriptsRecursively:
 
         with pytest.raises(ValueError) as e:
             get_all_scripts_recursively(Path("scripts"))
-        assert str(e.value).startswith(
-            "The script name A__initial.sql exists more than once (first_instance "
-        )
+        assert str(e.value).startswith("The script name A__initial.sql exists more than once (first_instance ")
 
     def test_given_single_always_file_should_extract_attributes(self, fs):
         fs.create_file(Path("scripts") / "subfolder" / "A__THIS_is_my_test.sql")
@@ -332,9 +304,7 @@ class TestGetAllScriptsRecursively:
         assert len(result) == 1
         script = result["a__this_is_my_test.sql"]
         assert script.name == "A__THIS_is_my_test.sql"
-        assert (
-            script.file_path == Path("scripts") / "subfolder" / "A__THIS_is_my_test.sql"
-        )
+        assert script.file_path == Path("scripts") / "subfolder" / "A__THIS_is_my_test.sql"
         assert script.type == "A"
         assert script.description == "This is my test"
 
@@ -345,24 +315,17 @@ class TestGetAllScriptsRecursively:
         assert len(result) == 1
         script = result["a__this_is_my_test.sql"]
         assert script.name == "A__THIS_is_my_test.sql"
-        assert (
-            script.file_path
-            == Path("scripts") / "subfolder" / "A__THIS_is_my_test.sql.jinja"
-        )
+        assert script.file_path == Path("scripts") / "subfolder" / "A__THIS_is_my_test.sql.jinja"
         assert script.type == "A"
         assert script.description == "This is my test"
 
-    def test_given_same_always_file_with_and_without_jinja_extension_should_raise_exception(
-        self, fs
-    ):
+    def test_given_same_always_file_with_and_without_jinja_extension_should_raise_exception(self, fs):
         fs.create_file(Path("scripts") / "A__initial.sql")
         fs.create_file(Path("scripts") / "A__initial.sql.jinja")
 
         with pytest.raises(ValueError) as e:
             get_all_scripts_recursively(Path("scripts"))
-        assert str(e.value).startswith(
-            "The script name A__initial.sql exists more than once (first_instance "
-        )
+        assert str(e.value).startswith("The script name A__initial.sql exists more than once (first_instance ")
 
     ###############################
     #### Repeatable file tests ####
@@ -384,9 +347,7 @@ class TestGetAllScriptsRecursively:
         fs.create_file(Path("scripts") / "subfolder" / "R__initial.SQL")
         with pytest.raises(ValueError) as e:
             get_all_scripts_recursively(Path("scripts"))
-        assert str(e.value).startswith(
-            "The script name R__initial.SQL exists more than once (first_instance "
-        )
+        assert str(e.value).startswith("The script name R__initial.SQL exists more than once (first_instance ")
 
     def test_given_single_repeatable_file_should_extract_attributes(self, fs):
         fs.create_file(Path("scripts") / "subfolder" / "R__THIS_is_my_test.sql")
@@ -395,9 +356,7 @@ class TestGetAllScriptsRecursively:
         assert len(result) == 1
         script = result["r__this_is_my_test.sql"]
         assert script.name == "R__THIS_is_my_test.sql"
-        assert (
-            script.file_path == Path("scripts") / "subfolder" / "R__THIS_is_my_test.sql"
-        )
+        assert script.file_path == Path("scripts") / "subfolder" / "R__THIS_is_my_test.sql"
         assert script.type == "R"
         assert script.description == "This is my test"
 
@@ -408,20 +367,13 @@ class TestGetAllScriptsRecursively:
         assert len(result) == 1
         script = result["r__this_is_my_test.sql"]
         assert script.name == "R__THIS_is_my_test.sql"
-        assert (
-            script.file_path
-            == Path("scripts") / "subfolder" / "R__THIS_is_my_test.sql.jinja"
-        )
+        assert script.file_path == Path("scripts") / "subfolder" / "R__THIS_is_my_test.sql.jinja"
         assert script.type == "R"
         assert script.description == "This is my test"
 
-    def test_given_same_repeatable_file_with_and_without_jinja_extension_should_raise_exception(
-        self, fs
-    ):
+    def test_given_same_repeatable_file_with_and_without_jinja_extension_should_raise_exception(self, fs):
         fs.create_file(Path("scripts") / "R__initial.sql")
         fs.create_file(Path("scripts") / "R__initial.sql.jinja")
         with pytest.raises(ValueError) as e:
             get_all_scripts_recursively(Path("scripts"))
-        assert str(e.value).startswith(
-            "The script name R__initial.sql exists more than once (first_instance "
-        )
+        assert str(e.value).startswith("The script name R__initial.sql exists more than once (first_instance ")
