@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-
 import pytest
 import structlog
 
@@ -725,14 +724,10 @@ class NoStr:
 class TestGetRedactConfigSecretsProcessor:
     @pytest.mark.parametrize("secrets, extra_kwargs, expected", cases)
     def test_happy_path(self, secrets: set[str], extra_kwargs: dict, expected: dict):
-        redact_config_secrets_processor = get_redact_config_secrets_processor(
-            config_secrets=secrets
-        )
+        redact_config_secrets_processor = get_redact_config_secrets_processor(config_secrets=secrets)
 
         # noinspection PyTypeChecker
-        result = redact_config_secrets_processor(
-            None, "info", {"event": "event text", "level": "info", **extra_kwargs}
-        )
+        result = redact_config_secrets_processor(None, "info", {"event": "event text", "level": "info", **extra_kwargs})
 
         assert result == {"event": "event text", "level": "info", **expected}
 
@@ -740,17 +735,7 @@ class TestGetRedactConfigSecretsProcessor:
         "extra_kwargs, expected_warning",
         [
             (
-                {
-                    "key_1": {
-                        "key_1": {
-                            "key_1": {
-                                "key_1": {
-                                    "key_1": {"key_1": {"key_1": {"key_1": "secret"}}}
-                                }
-                            }
-                        }
-                    }
-                },
+                {"key_1": {"key_1": {"key_1": {"key_1": {"key_1": {"key_1": {"key_1": {"key_1": "secret"}}}}}}}},
                 "Unable to redact deeply nested secrets in log",
             ),
             (
@@ -760,14 +745,10 @@ class TestGetRedactConfigSecretsProcessor:
         ],
     )
     def test_warnings(self, extra_kwargs: dict, expected_warning: str):
-        redact_config_secrets_processor = get_redact_config_secrets_processor(
-            config_secrets={"secret"}
-        )
+        redact_config_secrets_processor = get_redact_config_secrets_processor(config_secrets={"secret"})
 
         with pytest.warns(UserWarning) as e:
             # noinspection PyTypeChecker
-            redact_config_secrets_processor(
-                None, "info", {"event": "event text", "level": "info", **extra_kwargs}
-            )
+            redact_config_secrets_processor(None, "info", {"event": "event text", "level": "info", **extra_kwargs})
 
         assert expected_warning in str(e[0].message)
